# api/openapi.yaml
openapi: "3.1.0"
info:
  title: GoTP API
  version: "1.0.0"
  description: >
    OTP-as-a-Service over SMS with a built-in fraud guard. Before any code is
    sent, four auditable checks run — per-number/IP velocity, send-vs-verify
    ratio per prefix, a fan-out (spray) detector, and an optional SIM-swap
    check. Verification is hardened with hashed codes, expiry, attempt lockout,
    and atomic check-and-consume.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://{instance}.africastalking.marketplace.dev
    description: Marketplace instance
    variables:
      instance:
        default: your-app
...
paths:
...
  /debug/vars:
    get:
      tags: [System]
      summary: Runtime metrics (expvar)
      description: >
        expvar metrics including total_requests_received, total_responses_sent,
        blocked_sends_total, and per-status response tallies.
      operationId: metrics
      security:
        - {}
      responses:
        "200":
          description: expvar JSON payload.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Required only when the instance is configured with API keys. The key is
        compared by SHA-256 hash; the plaintext is never stored.
...